refactor: `pagination.js` => `pagination-bar.js`
