+++ src/script/components/pagination-bar.js
